fix all issues with .NET Core
