+++ dotnet/Gherkin.CSharp.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gherkin.Specs", "Gherkin.Sp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gherkin.TokensGenerator", "Gherkin.TokensGenerator\Gherkin.TokensGenerator.csproj", "{4725D7DF-7488-41CC-8F30-90C6A5FB463A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Gherkin.TokensGeneratorLib", "Gherkin.TokensGeneratorLib\Gherkin.TokensGeneratorLib.csproj", "{77918F35-3CB3-4A62-95F9-438804A2BF02}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -69,6 +71,18 @@ Global
 		{4725D7DF-7488-41CC-8F30-90C6A5FB463A}.Release|x64.Build.0 = Release|x64
 		{4725D7DF-7488-41CC-8F30-90C6A5FB463A}.Release|x86.ActiveCfg = Release|x86
 		{4725D7DF-7488-41CC-8F30-90C6A5FB463A}.Release|x86.Build.0 = Release|x86
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Debug|x64.ActiveCfg = Debug|x64
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Debug|x64.Build.0 = Debug|x64
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Debug|x86.ActiveCfg = Debug|x86
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Debug|x86.Build.0 = Debug|x86
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Release|Any CPU.Build.0 = Release|Any CPU
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Release|x64.ActiveCfg = Release|x64
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Release|x64.Build.0 = Release|x64
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Release|x86.ActiveCfg = Release|x86
+		{77918F35-3CB3-4A62-95F9-438804A2BF02}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
