+++ dotnet/Gherkin.TokensGeneratorLib/Gherkin.TokensGeneratorLib.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>netstandard1.3;net45</TargetFrameworks>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Gherkin\Gherkin.csproj" />
+  </ItemGroup>
+
+</Project>
