--- dotnet/global.json
@@ -1,7 +1,0 @@
-{
-  "projects": [ "Gherkin" ],
-  "sdk": {
-    "version": "1.0.0-preview2-003156"
-  }
-}
-
